colab_requirements.txt
# Core
flask
flask-cors
python-dotenv
pyngrok

# Download de Vídeos
yt-dlp>=2024.0.0

# Transcrição & AI
faster-whisper
langchain-ollama
torch
transformers
sentencepiece
vosk
# edge-tts
# onnxruntime

# Processamento de Vídeo
moviepy==1.0.3
ffmpeg-python
opencv-python-headless
Pillow

# Áudio
librosa
pydub
noisereduce
soundfile
edge-tts
soundfile

# NLP
nltk
spacy

# Utils
requests
selenium
tqdm
numpy
pandas
